add support for multiple signing tools
