+++ cmd/timoni/mod_push.go
@@ -79,7 +79,7 @@ type pushModFlags struct {
 	ignorePaths []string
 	output      string
 	annotations []string
-	sign        bool
+	sign        string
 	cosignKey   string
 }
 
@@ -96,8 +96,8 @@ func init() {
 		"Set custom OCI annotations in the format '<key>=<value>'.")
 	pushModCmd.Flags().StringVarP(&pushModArgs.output, "output", "o", "",
 		"The format in which the artifact digest should be printed, can be 'yaml' or 'json'.")
-	pushModCmd.Flags().BoolVar(&pushModArgs.sign, "sign", false,
-		"Signs the module with Cosign.")
+	pushModCmd.Flags().StringVar(&pushModArgs.sign, "sign", "",
+		"Signs the module with the specified provider.")
 	pushModCmd.Flags().StringVar(&pushModArgs.cosignKey, "cosign-key", "",
 		"The Cosign private key for signing the module.")
 
@@ -201,8 +201,8 @@ func pushModCmdRun(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("artifact tag parsing failed: %w", err)
 	}
 
-	if pushModArgs.sign {
-		err = signutil.Sign(log, digestURL, pushModArgs.cosignKey)
+	if pushModArgs.sign != "" {
+		err = signutil.Sign(log, pushModArgs.sign, digestURL, pushModArgs.cosignKey)
 		if err != nil {
 			return err
 		}
